moved examples to their respective crates
