+++ Cargo.toml
@@ -5,5 +5,3 @@ members = [
 ]
 
 resolver = "2"
-
-
